dont remove environment if group uses it

+++ src/Controllers/Environment/RemoveEnvironmentController.php
@@ -8,7 +8,9 @@
 namespace Hal\UI\Controllers\Environment;
 
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\EntityRepository;
 use Hal\Core\Entity\Environment;
+use Hal\Core\Entity\Group;
 use Hal\UI\Controllers\RedirectableControllerTrait;
 use Hal\UI\Controllers\SessionTrait;
 use Hal\UI\Flash;
@@ -23,7 +25,7 @@ class RemoveEnvironmentController implements ControllerInterface
     use SessionTrait;
 
     private const MSG_SUCCESS = '"%s" environment removed.';
-    // private const ERR_HAS_SERVERS = 'Cannot remove environment. All associated servers must first be removed.';
+    private const ERR_HAS_GROUPS = 'Cannot remove environment. All associated groups must first be removed.';
 
     /**
      * @var EntityManagerInterface
@@ -36,12 +38,17 @@ class RemoveEnvironmentController implements ControllerInterface
     private $uri;
 
     /**
+     * @var EntityRepository
+     */
+    private $groupRepository;
+
+    /**
      * @param EntityManagerInterface $em
      * @param URI $uri
      */
     public function __construct(EntityManagerInterface $em, URI $uri)
     {
-        // $this->serverRepo = $em->getRepository(Server::class);
+        $this->groupRepository = $em->getRepository(Group::class);
         $this->em = $em;
         $this->uri = $uri;
     }
@@ -53,10 +60,10 @@ class RemoveEnvironmentController implements ControllerInterface
     {
         $environment = $request->getAttribute(Environment::class);
 
-        // if ($servers = $this->serverRepo->findBy(['environment' => $environment])) {
-        //     $this->withFlash($request, Flash::ERROR, self::ERR_HAS_SERVERS);
-        //     return $this->withRedirectRoute($response, $this->uri, 'environment', ['environment' => $environment->id()]);
-        // }
+         if ($groups = $this->groupRepository->findBy(['environment' => $environment])) {
+             $this->withFlash($request, Flash::ERROR, self::ERR_HAS_GROUPS);
+             return $this->withRedirectRoute($response, $this->uri, 'environment', ['environment' => $environment->id()]);
+         }
 
         $this->em->remove($environment);
         $this->em->flush();
